Move function to common utilities
