+++ lib/tabs/utils.js
@@ -19,20 +19,3 @@ export function getTabElementId(tabInterfaceId, tabId) {
 export function getTabPanelElementId(tabInterfaceId, tabId) {
   return getTabElementId(tabInterfaceId, tabId) + '-panel';
 }
-
-/**
- * @param {HTMLElement} parent A parent element
- * @returns {HTMLElement, null} The first focusable child of the parent
- *                              element or `null` when none is found
- */
-export function getFirstFocusableChild(parent) {
-  if (!parent) {
-    console.error(`[getFirstFocusableChild] 'parent' parameter is required`);
-    return null;
-  }
-  // https://gomakethings.com/how-to-get-the-first-and-last-focusable-elements-in-the-dom/
-  const focusableElements = parent.querySelectorAll(
-    'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])'
-  );
-  return focusableElements.length > 0 ? focusableElements[0] : null;
-}
